+++ src/lib/mlxcel-core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mlxcel-core"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 description = "Low-level MLX building blocks for mlxcel"
 license = "Apache-2.0"
