+++ src/lib/mlxcel-core/cpp/mlx_cxx_bridge.cpp
@@ -1,101 +1,14 @@
 // Copyright 2025 mlx-lm-rs authors
 // Direct C++ bridge implementation for MLX via cxx
 
-#include "mlx_cxx_bridge.h"
-#include "sparse_v_sdpa.h"  // fused Sparse-V SDPA kernel.
-#include "turbo4_delegated_sdpa.h"  // fused Turbo4Delegated SDPA kernel.
-#include "paged_attention.h"  // fused paged-attention decode kernel (#123).
-#include "mlx/ops.h"
-#include "mlx/transforms.h"
-#include "mlx/compile.h"
-#include "mlx/memory.h"
-#include "mlx/linalg.h"
-#include "mlx/fft.h"
-#include "mlx/graph_utils.h"
-#include "mlx/random.h"
-#include "mlx/einsum.h"
-#include "mlx/utils.h"
-#ifdef __APPLE__
-#include "mlx/backend/metal/metal.h"
-#endif
-#include <algorithm>
-#include <cstring>
-#include <cstdlib>
-#include <fstream>
-#include <iostream>
-#include <mutex>
-#include <stdexcept>
-#include <unordered_map>
+#include "mlx_cxx_internal.h"
 
 namespace mlx_cxx {
 
 using namespace mlx::core;
 
-// Helper to convert rust slice to std::vector<int32_t> (Shape)
-static Shape to_shape(rust::Slice<const int32_t> slice) {
-    return Shape(slice.begin(), slice.end());
-}
-
-// Helper to convert Dtype int to mlx::core::Dtype
-static Dtype to_dtype(int32_t dtype) {
-    switch (dtype) {
-        case 0: return bool_;
-        case 1: return uint8;
-        case 2: return uint16;
-        case 3: return uint32;
-        case 4: return uint64;
-        case 5: return int8;
-        case 6: return int16;
-        case 7: return int32;
-        case 8: return int64;
-        case 9: return float16;
-        case 10: return float32;
-        case 11: return float64;
-        case 12: return bfloat16;
-        case 13: return complex64;
-        default: return float32;
-    }
-}
-
-// Helper to convert mlx::core::Dtype to int
-static int32_t from_dtype(Dtype dtype) {
-    switch (dtype.val()) {
-        case bool_.val(): return 0;
-        case uint8.val(): return 1;
-        case uint16.val(): return 2;
-        case uint32.val(): return 3;
-        case uint64.val(): return 4;
-        case int8.val(): return 5;
-        case int16.val(): return 6;
-        case int32.val(): return 7;
-        case int64.val(): return 8;
-        case float16.val(): return 9;
-        case float32.val(): return 10;
-        case float64.val(): return 11;
-        case bfloat16.val(): return 12;
-        case complex64.val(): return 13;
-        default: return 10;  // float32
-    }
-}
-
-// MLX Python's nn.gelu_approx uses the tanh approximation:
-// 0.5 * x * (1 + tanh(sqrt(2/pi) * (x + 0.044715 * x^3))).
-// Keep this helper local so exact-GELU models can continue to use the
-// erf-based compiled_geglu_activation path.
-static array gelu_tanh_approx(const array& x) {
-    auto dtype = x.dtype();
-    auto half = array(0.5f, dtype);
-    auto one = array(1.0f, dtype);
-    auto coeff = array(0.7978845608028654f, dtype);  // sqrt(2/pi)
-    auto cubic_coeff = array(0.044715f, dtype);
-    auto x2 = mlx::core::multiply(x, x);
-    auto x3 = mlx::core::multiply(x2, x);
-    auto inner = mlx::core::multiply(
-        coeff,
-        mlx::core::add(x, mlx::core::multiply(cubic_coeff, x3)));
-    auto cdf = mlx::core::multiply(half, mlx::core::add(one, mlx::core::tanh(inner)));
-    return mlx::core::multiply(x, cdf);
-}
+// to_shape / to_dtype / from_dtype / gelu_tanh_approx live in
+// mlx_cxx_internal.h so the split-out TUs (kernels/nemotron/ext) share them.
 
 // Stream functions.
 std::unique_ptr<MlxStream> default_stream() {
@@ -4700,1668 +4613,4 @@ std::unique_ptr<MlxArray> quantize_weights_biases(const MlxArray& w, int32_t gro
     return std::make_unique<MlxArray>(std::move(result[2]));
 }
 
-// SSM (Mamba2) fused Metal kernel for single-token decode.
-// Port of Python mlx-lm ssm.py make_ssm_kernel() + ssm_update_kernel()
-namespace {
-    static const char* SSM_METAL_SOURCE = R"(
-        auto n = thread_position_in_grid.z;
-        auto h_idx = n % H;
-        auto g_idx = n / G;
-        constexpr int n_per_t = Ds / 32;
-
-        auto x = X + n * Dh;
-        out += n * Dh;
-        auto i_state = state_in + n * Dh * Ds;
-        auto o_state = state_out + n * Dh * Ds;
-
-        auto C_ = C + g_idx * Ds;
-        auto B_ = B + g_idx * Ds;
-
-        auto ds_idx = thread_position_in_threadgroup.x;
-        auto d_idx = thread_position_in_grid.y;
-
-        auto dt_ = static_cast<float>(dt[n]);
-        auto A = -fast::exp(static_cast<float>(A_log[h_idx]));
-        auto dA = fast::exp(A * dt_);
-
-        float acc = 0.0;
-        auto x_ = static_cast<float>(x[d_idx]);
-
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * ds_idx + i;
-            auto idx = d_idx * Ds + s_idx;
-            auto dB_by_x = x_ * dt_ * static_cast<float>(B_[s_idx]);
-            auto state = dA * i_state[idx] + dB_by_x;
-            o_state[idx] = static_cast<U>(state);
-            acc += state * C_[s_idx];
-        }
-        acc = simd_sum(acc);
-        if (thread_index_in_simdgroup == 0) {
-            out[d_idx] = static_cast<T>(acc + x_ * D[h_idx]);
-        }
-    )";
-
-    struct SsmKernelHolder {
-        std::optional<mlx::core::fast::CustomKernelFunction> kernel;
-        bool initialized = false;
-
-        mlx::core::fast::CustomKernelFunction& get() {
-            if (!initialized) {
-                kernel = mlx::core::fast::metal_kernel(
-                    "ssm_kernel",
-                    {"X", "A_log", "B", "C", "D", "dt", "state_in"},
-                    {"out", "state_out"},
-                    SSM_METAL_SOURCE
-                );
-                initialized = true;
-            }
-            return *kernel;
-        }
-    };
-
-    static SsmKernelHolder& get_ssm_kernel() {
-        static SsmKernelHolder holder;
-        return holder;
-    }
-
-    // Compiled compute_dt: float32 promotion + softplus + clip → single fused kernel
-    // Matches Python's @mx.compile compute_dt (casts dt to float32 before softplus for precision)
-    static std::function<std::vector<array>(const std::vector<array>&)>
-    get_compiled_compute_dt() {
-        auto fn = [](const std::vector<array>& inputs) -> std::vector<array> {
-            auto dt = mlx::core::astype(inputs[0], mlx::core::float32);
-            const auto& dt_bias = inputs[1];
-            const auto& lo = inputs[2];
-            const auto& hi = inputs[3];
-            auto result = mlx::core::add(dt, dt_bias);
-            result = mlx::core::log1p(mlx::core::exp(result));
-            return {mlx::core::clip(result, lo, hi)};
-        };
-        return mlx::core::compile(fn, true);
-    }
-
-    static array compute_dt_compiled(const array& dt, const array& dt_bias, float min_val, float max_val) {
-        static auto compiled_fn = get_compiled_compute_dt();
-        auto lo = mlx::core::array(min_val);
-        auto hi = mlx::core::array(max_val);
-        return compiled_fn({dt, dt_bias, lo, hi})[0];
-    }
-}
-
-// GatedDeltaNet custom Metal kernel for single/multi-token decode.
-// Port of Python mlx-lm gated_delta.py _make_gated_delta_kernel()
-// Four variants: scalar/vec gate x mask/no-mask
-namespace {
-    // Variant 1: scalar gate, no mask
-    static const char* GATED_DELTA_METAL_SOURCE = R"(
-        auto n = thread_position_in_grid.z;
-        auto b_idx = n / Hv;
-        auto hv_idx = n % Hv;
-        auto hk_idx = hv_idx / (Hv / Hk);
-        constexpr int n_per_t = Dk / 32;
-
-        // q, k: [B, T, Hk, Dk]
-        auto q_ = q + b_idx * T * Hk * Dk + hk_idx * Dk;
-        auto k_ = k + b_idx * T * Hk * Dk + hk_idx * Dk;
-
-        // v, y: [B, T, Hv, Dv]
-        auto v_ = v + b_idx * T * Hv * Dv + hv_idx * Dv;
-        y += b_idx * T * Hv * Dv + hv_idx * Dv;
-
-        auto dk_idx = thread_position_in_threadgroup.x;
-        auto dv_idx = thread_position_in_grid.y;
-
-        // state_in, state_out: [B, Hv, Dv, Dk]
-        auto i_state = state_in + (n * Dv + dv_idx) * Dk;
-        auto o_state = state_out + (n * Dv + dv_idx) * Dk;
-
-        float state[n_per_t];
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * dk_idx + i;
-            state[i] = static_cast<float>(i_state[s_idx]);
-        }
-
-        // g: [B, T, Hv]
-        auto g_ = g + b_idx * T * Hv;
-        auto beta_ = beta + b_idx * T * Hv;
-
-        for (int t = 0; t < T; ++t) {
-            if (true) {
-                float kv_mem = 0.0f;
-                for (int i = 0; i < n_per_t; ++i) {
-                    auto s_idx = n_per_t * dk_idx + i;
-                    state[i] = state[i] * g_[hv_idx];
-                    kv_mem += state[i] * k_[s_idx];
-                }
-                kv_mem = simd_sum(kv_mem);
-
-                auto delta = (v_[dv_idx] - kv_mem) * beta_[hv_idx];
-
-                float out = 0.0f;
-                for (int i = 0; i < n_per_t; ++i) {
-                    auto s_idx = n_per_t * dk_idx + i;
-                    state[i] = state[i] + k_[s_idx] * delta;
-                    out += state[i] * q_[s_idx];
-                }
-                out = simd_sum(out);
-                if (thread_index_in_simdgroup == 0) {
-                    y[dv_idx] = static_cast<InT>(out);
-                }
-            }
-            q_ += Hk * Dk;
-            k_ += Hk * Dk;
-            v_ += Hv * Dv;
-            y += Hv * Dv;
-            g_ += Hv;
-            beta_ += Hv;
-        }
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * dk_idx + i;
-            o_state[s_idx] = static_cast<InT>(state[i]);
-        }
-    )";
-
-    // Variant 2: scalar gate, with mask
-    static const char* GATED_DELTA_METAL_SOURCE_MASK = R"(
-        auto n = thread_position_in_grid.z;
-        auto b_idx = n / Hv;
-        auto hv_idx = n % Hv;
-        auto hk_idx = hv_idx / (Hv / Hk);
-        constexpr int n_per_t = Dk / 32;
-
-        auto q_ = q + b_idx * T * Hk * Dk + hk_idx * Dk;
-        auto k_ = k + b_idx * T * Hk * Dk + hk_idx * Dk;
-
-        auto v_ = v + b_idx * T * Hv * Dv + hv_idx * Dv;
-        y += b_idx * T * Hv * Dv + hv_idx * Dv;
-
-        auto dk_idx = thread_position_in_threadgroup.x;
-        auto dv_idx = thread_position_in_grid.y;
-
-        auto i_state = state_in + (n * Dv + dv_idx) * Dk;
-        auto o_state = state_out + (n * Dv + dv_idx) * Dk;
-
-        float state[n_per_t];
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * dk_idx + i;
-            state[i] = static_cast<float>(i_state[s_idx]);
-        }
-
-        // g: [B, T, Hv]
-        auto g_ = g + b_idx * T * Hv;
-        auto beta_ = beta + b_idx * T * Hv;
-
-        for (int t = 0; t < T; ++t) {
-            if (mask[b_idx * T + t]) {
-                float kv_mem = 0.0f;
-                for (int i = 0; i < n_per_t; ++i) {
-                    auto s_idx = n_per_t * dk_idx + i;
-                    state[i] = state[i] * g_[hv_idx];
-                    kv_mem += state[i] * k_[s_idx];
-                }
-                kv_mem = simd_sum(kv_mem);
-
-                auto delta = (v_[dv_idx] - kv_mem) * beta_[hv_idx];
-
-                float out = 0.0f;
-                for (int i = 0; i < n_per_t; ++i) {
-                    auto s_idx = n_per_t * dk_idx + i;
-                    state[i] = state[i] + k_[s_idx] * delta;
-                    out += state[i] * q_[s_idx];
-                }
-                out = simd_sum(out);
-                if (thread_index_in_simdgroup == 0) {
-                    y[dv_idx] = static_cast<InT>(out);
-                }
-            } else {
-                y[dv_idx] = static_cast<InT>(0);
-            }
-            q_ += Hk * Dk;
-            k_ += Hk * Dk;
-            v_ += Hv * Dv;
-            y += Hv * Dv;
-            g_ += Hv;
-            beta_ += Hv;
-        }
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * dk_idx + i;
-            o_state[s_idx] = static_cast<InT>(state[i]);
-        }
-    )";
-
-    // Variant 3: vectorized gate (per-dim), no mask
-    static const char* GATED_DELTA_METAL_SOURCE_VEC = R"(
-        auto n = thread_position_in_grid.z;
-        auto b_idx = n / Hv;
-        auto hv_idx = n % Hv;
-        auto hk_idx = hv_idx / (Hv / Hk);
-        constexpr int n_per_t = Dk / 32;
-
-        auto q_ = q + b_idx * T * Hk * Dk + hk_idx * Dk;
-        auto k_ = k + b_idx * T * Hk * Dk + hk_idx * Dk;
-
-        auto v_ = v + b_idx * T * Hv * Dv + hv_idx * Dv;
-        y += b_idx * T * Hv * Dv + hv_idx * Dv;
-
-        auto dk_idx = thread_position_in_threadgroup.x;
-        auto dv_idx = thread_position_in_grid.y;
-
-        auto i_state = state_in + (n * Dv + dv_idx) * Dk;
-        auto o_state = state_out + (n * Dv + dv_idx) * Dk;
-
-        float state[n_per_t];
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * dk_idx + i;
-            state[i] = static_cast<float>(i_state[s_idx]);
-        }
-
-        // g: [B, T, Hv, Dk]
-        auto g_ = g + (b_idx * T * Hv + hv_idx) * Dk;
-        auto beta_ = beta + b_idx * T * Hv;
-
-        for (int t = 0; t < T; ++t) {
-            if (true) {
-                float kv_mem = 0.0f;
-                for (int i = 0; i < n_per_t; ++i) {
-                    auto s_idx = n_per_t * dk_idx + i;
-                    state[i] = state[i] * g_[s_idx];
-                    kv_mem += state[i] * k_[s_idx];
-                }
-                kv_mem = simd_sum(kv_mem);
-
-                auto delta = (v_[dv_idx] - kv_mem) * beta_[hv_idx];
-
-                float out = 0.0f;
-                for (int i = 0; i < n_per_t; ++i) {
-                    auto s_idx = n_per_t * dk_idx + i;
-                    state[i] = state[i] + k_[s_idx] * delta;
-                    out += state[i] * q_[s_idx];
-                }
-                out = simd_sum(out);
-                if (thread_index_in_simdgroup == 0) {
-                    y[dv_idx] = static_cast<InT>(out);
-                }
-            }
-            q_ += Hk * Dk;
-            k_ += Hk * Dk;
-            v_ += Hv * Dv;
-            y += Hv * Dv;
-            g_ += Hv * Dk;
-            beta_ += Hv;
-        }
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * dk_idx + i;
-            o_state[s_idx] = static_cast<InT>(state[i]);
-        }
-    )";
-
-    // Variant 4: vectorized gate, with mask
-    static const char* GATED_DELTA_METAL_SOURCE_VEC_MASK = R"(
-        auto n = thread_position_in_grid.z;
-        auto b_idx = n / Hv;
-        auto hv_idx = n % Hv;
-        auto hk_idx = hv_idx / (Hv / Hk);
-        constexpr int n_per_t = Dk / 32;
-
-        auto q_ = q + b_idx * T * Hk * Dk + hk_idx * Dk;
-        auto k_ = k + b_idx * T * Hk * Dk + hk_idx * Dk;
-
-        auto v_ = v + b_idx * T * Hv * Dv + hv_idx * Dv;
-        y += b_idx * T * Hv * Dv + hv_idx * Dv;
-
-        auto dk_idx = thread_position_in_threadgroup.x;
-        auto dv_idx = thread_position_in_grid.y;
-
-        auto i_state = state_in + (n * Dv + dv_idx) * Dk;
-        auto o_state = state_out + (n * Dv + dv_idx) * Dk;
-
-        float state[n_per_t];
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * dk_idx + i;
-            state[i] = static_cast<float>(i_state[s_idx]);
-        }
-
-        // g: [B, T, Hv, Dk]
-        auto g_ = g + (b_idx * T * Hv + hv_idx) * Dk;
-        auto beta_ = beta + b_idx * T * Hv;
-
-        for (int t = 0; t < T; ++t) {
-            if (mask[b_idx * T + t]) {
-                float kv_mem = 0.0f;
-                for (int i = 0; i < n_per_t; ++i) {
-                    auto s_idx = n_per_t * dk_idx + i;
-                    state[i] = state[i] * g_[s_idx];
-                    kv_mem += state[i] * k_[s_idx];
-                }
-                kv_mem = simd_sum(kv_mem);
-
-                auto delta = (v_[dv_idx] - kv_mem) * beta_[hv_idx];
-
-                float out = 0.0f;
-                for (int i = 0; i < n_per_t; ++i) {
-                    auto s_idx = n_per_t * dk_idx + i;
-                    state[i] = state[i] + k_[s_idx] * delta;
-                    out += state[i] * q_[s_idx];
-                }
-                out = simd_sum(out);
-                if (thread_index_in_simdgroup == 0) {
-                    y[dv_idx] = static_cast<InT>(out);
-                }
-            } else {
-                y[dv_idx] = static_cast<InT>(0);
-            }
-            q_ += Hk * Dk;
-            k_ += Hk * Dk;
-            v_ += Hv * Dv;
-            y += Hv * Dv;
-            g_ += Hv * Dk;
-            beta_ += Hv;
-        }
-        for (int i = 0; i < n_per_t; ++i) {
-            auto s_idx = n_per_t * dk_idx + i;
-            o_state[s_idx] = static_cast<InT>(state[i]);
-        }
-    )";
-
-    // Kernel holder structs (lazy init, one per variant)
-    struct GatedDeltaKernelHolder {
-        std::optional<mlx::core::fast::CustomKernelFunction> kernel;
-        bool initialized = false;
-
-        mlx::core::fast::CustomKernelFunction& get(const char* name,
-                                                    const std::vector<std::string>& inputs,
-                                                    const char* source) {
-            if (!initialized) {
-                kernel = mlx::core::fast::metal_kernel(
-                    name, inputs,
-                    {"y", "state_out"},
-                    source
-                );
-                initialized = true;
-            }
-            return *kernel;
-        }
-    };
-
-    static GatedDeltaKernelHolder& get_gd_kernel() {
-        static GatedDeltaKernelHolder holder;
-        return holder;
-    }
-    static GatedDeltaKernelHolder& get_gd_kernel_mask() {
-        static GatedDeltaKernelHolder holder;
-        return holder;
-    }
-    static GatedDeltaKernelHolder& get_gd_kernel_vec() {
-        static GatedDeltaKernelHolder holder;
-        return holder;
-    }
-    static GatedDeltaKernelHolder& get_gd_kernel_vec_mask() {
-        static GatedDeltaKernelHolder holder;
-        return holder;
-    }
-}
-
-bool gated_delta_kernel_available() {
-#ifdef __APPLE__
-    return mlx::core::metal::is_available();
-#else
-    return false;
-#endif
-}
-
-// Start a GPU trace capture and stop an active one. Mirrors the Python
-// `mx.metal.start_capture` / `mx.metal.stop_capture` API so a mlxcel
-// decode run can emit the same `.gputrace` files that `mlx-lm` produces
-// and both be loaded into Xcode's Metal Debugger side by side.
-// The process must have been launched with `MTL_CAPTURE_ENABLED=1`,
-// otherwise Metal silently drops the capture.
-void metal_start_capture(rust::Str path) {
-    mlx::core::metal::start_capture(std::string(path.data(), path.size()));
-}
-
-void metal_stop_capture() {
-    mlx::core::metal::stop_capture();
-}
-
-void metal_gated_delta_forward(
-    const MlxArray& q,
-    const MlxArray& k,
-    const MlxArray& v,
-    const MlxArray& g,
-    const MlxArray& beta,
-    const MlxArray& state,
-    const MlxArray* mask,
-    std::unique_ptr<MlxArray>& output,
-    std::unique_ptr<MlxArray>& new_state
-) {
-    using namespace mlx::core;
-
-    // Extract dimensions from input shapes
-    auto q_shape = q.inner.shape();
-    int B = q_shape[0];
-    int T_val = q_shape[1];
-    int Hk = q_shape[2];
-    int Dk = q_shape[3];
-    auto v_shape = v.inner.shape();
-    int Hv = v_shape[2];
-    int Dv = v_shape[3];
-
-    auto input_type = q.inner.dtype();
-
-    // Cast inputs to input_type if needed (state may be float32 from Rust side)
-    auto state_cast = (state.inner.dtype() != input_type)
-        ? mlx::core::astype(state.inner, input_type) : state.inner;
-    auto g_cast = (g.inner.dtype() != input_type)
-        ? mlx::core::astype(g.inner, input_type) : g.inner;
-    auto beta_cast = (beta.inner.dtype() != input_type)
-        ? mlx::core::astype(beta.inner, input_type) : beta.inner;
-
-    // Detect vectorized gate: g has 4 dims [B, T, Hv, Dk]
-    bool vectorized = (g.inner.ndim() == 4);
-    bool has_mask = (mask != nullptr);
-
-    // Build T as a scalar array input
-    auto T_arr = mlx::core::array(T_val);
-
-    // Select kernel variant and build inputs
-    std::vector<array> inputs;
-    GatedDeltaKernelHolder* holder;
-    const char* kernel_name;
-    const char* kernel_source;
-    std::vector<std::string> input_names;
-
-    if (vectorized && has_mask) {
-        input_names = {"q", "k", "v", "g", "beta", "state_in", "T", "mask"};
-        inputs = {q.inner, k.inner, v.inner, g_cast, beta_cast, state_cast, T_arr, mask->inner};
-        holder = &get_gd_kernel_vec_mask();
-        kernel_name = "gated_delta_step_vec_mask";
-        kernel_source = GATED_DELTA_METAL_SOURCE_VEC_MASK;
-    } else if (vectorized) {
-        input_names = {"q", "k", "v", "g", "beta", "state_in", "T"};
-        inputs = {q.inner, k.inner, v.inner, g_cast, beta_cast, state_cast, T_arr};
-        holder = &get_gd_kernel_vec();
-        kernel_name = "gated_delta_step_vec";
-        kernel_source = GATED_DELTA_METAL_SOURCE_VEC;
-    } else if (has_mask) {
-        input_names = {"q", "k", "v", "g", "beta", "state_in", "T", "mask"};
-        inputs = {q.inner, k.inner, v.inner, g_cast, beta_cast, state_cast, T_arr, mask->inner};
-        holder = &get_gd_kernel_mask();
-        kernel_name = "gated_delta_step_mask";
-        kernel_source = GATED_DELTA_METAL_SOURCE_MASK;
-    } else {
-        input_names = {"q", "k", "v", "g", "beta", "state_in", "T"};
-        inputs = {q.inner, k.inner, v.inner, g_cast, beta_cast, state_cast, T_arr};
-        holder = &get_gd_kernel();
-        kernel_name = "gated_delta_step";
-        kernel_source = GATED_DELTA_METAL_SOURCE;
-    }
-
-    auto& kernel = holder->get(kernel_name, input_names, kernel_source);
-
-    // Template parameters: InT (dtype), Dk, Dv, Hk, Hv
-    std::vector<std::pair<std::string, mlx::core::fast::TemplateArg>> template_args = {
-        {"InT", input_type},
-        {"Dk", Dk},
-        {"Dv", Dv},
-        {"Hk", Hk},
-        {"Hv", Hv},
-    };
-
-    // Output shapes and dtypes (matching Python: both in input_type)
-    std::vector<Shape> output_shapes = {
-        Shape{B, T_val, Hv, Dv},   // y
-        state.inner.shape(),        // state_out (same shape as state_in)
-    };
-    std::vector<Dtype> output_dtypes = {input_type, input_type};
-
-    // Grid: (32, Dv, B * Hv), Threadgroup: (32, 4, 1)
-    auto results = kernel(
-        inputs,
-        output_shapes,
-        output_dtypes,
-        std::make_tuple(32, Dv, B * Hv),   // grid
-        std::make_tuple(32, 4, 1),          // threadgroup
-        template_args,
-        std::nullopt,  // init_value
-        false,         // verbose
-        {}             // stream (default)
-    );
-
-    output = std::make_unique<MlxArray>(std::move(results[0]));
-    new_state = std::make_unique<MlxArray>(std::move(results[1]));
-}
-
-// Compiled MoE gate: sigmoid + bias + topk + normalize + scale
-// Uses mx::core::compile for kernel fusion matching Python's @mx.compile group_expert_select
-namespace {
-    // Pre-sigmoid + bias + negative: compiled into fused kernel
-    static std::function<std::vector<array>(const std::vector<array>&)>
-    get_compiled_gate_scores() {
-        auto fn = [](const std::vector<array>& inputs) -> std::vector<array> {
-            const auto& gates = inputs[0];
-            const auto& bias = inputs[1];
-            auto orig = mlx::core::sigmoid(mlx::core::astype(gates, mlx::core::float32));
-            auto biased = mlx::core::add(orig, bias);
-            auto neg = mlx::core::negative(biased);
-            return {orig, neg};
-        };
-        return mlx::core::compile(fn, true);
-    }
-
-    // Post-topk normalize + scale: compiled into fused kernel
-    static std::function<std::vector<array>(const std::vector<array>&)>
-    get_compiled_gate_normalize() {
-        auto fn = [](const std::vector<array>& inputs) -> std::vector<array> {
-            const auto& scores = inputs[0];
-            const auto& scale = inputs[1];
-            auto denom = mlx::core::sum(scores, -1, true);
-            auto normed = mlx::core::divide(scores, mlx::core::add(denom, mlx::core::array(1e-20f)));
-            return {mlx::core::multiply(normed, scale)};
-        };
-        return mlx::core::compile(fn, true);
-    }
-}
-
-void compiled_moe_gate(
-    const MlxArray& gates,
-    const MlxArray& correction_bias,
-    int32_t top_k,
-    float scaling_factor,
-    bool norm_topk_prob,
-    std::unique_ptr<MlxArray>& indices_out,
-    std::unique_ptr<MlxArray>& scores_out
-) {
-    using namespace mlx::core;
-
-    // Step 1: compiled sigmoid + bias + negative
-    static auto gate_scores_fn = get_compiled_gate_scores();
-    auto gate_results = gate_scores_fn({gates.inner, correction_bias.inner});
-    auto orig_scores = gate_results[0];  // sigmoid(float32)
-    auto neg_biased = gate_results[1];   // negative(sigmoid + bias)
-
-    // Step 2: argpartition + slice (not compilable due to dynamic shapes)
-    auto indices = argpartition(neg_biased, top_k - 1, -1);
-    auto topk_indices = slice(indices, {0, 0}, {(int)indices.shape()[0], top_k});
-    auto topk_scores = take_along_axis(orig_scores, topk_indices, -1);
-
-    // Step 3: compiled normalize + scale
-    if (top_k > 1 && norm_topk_prob) {
-        static auto normalize_fn = get_compiled_gate_normalize();
-        auto scale = array(scaling_factor);
-        auto norm_results = normalize_fn({topk_scores, scale});
-        topk_scores = norm_results[0];
-    } else {
-        topk_scores = multiply(topk_scores, array(scaling_factor));
-    }
-
-    indices_out = std::make_unique<MlxArray>(std::move(topk_indices));
-    scores_out = std::make_unique<MlxArray>(std::move(topk_scores));
-}
-
-// Fused MoE forward: combines gate + switch_mlp + score weighting + shared expert
-std::unique_ptr<MlxArray> fused_moe_forward(
-    const MlxArray& x,
-    const MlxArray& gate_weight,
-    const MlxArray& correction_bias,
-    const MlxArray& fc1_weight,
-    const MlxArray& fc1_scales,
-    const MlxArray& fc1_biases,
-    const MlxArray& fc2_weight,
-    const MlxArray& fc2_scales,
-    const MlxArray& fc2_biases,
-    const MlxArray* shared_up_weight,
-    const MlxArray* shared_up_scales,
-    const MlxArray* shared_up_biases,
-    const MlxArray* shared_down_weight,
-    const MlxArray* shared_down_scales,
-    const MlxArray* shared_down_biases,
-    int32_t top_k,
-    float scaling_factor,
-    bool norm_topk_prob,
-    int32_t group_size,
-    int32_t bits
-) {
-    using namespace mlx::core;
-
-    // 1. Gate: compiled sigmoid + topk + compiled normalize + scale
-    auto gates = matmul(x.inner, transpose(gate_weight.inner));
-
-    // Compiled: sigmoid(astype(gates, f32)) + add(bias) + negative
-    static auto gate_scores_fn = get_compiled_gate_scores();
-    auto gate_results = gate_scores_fn({gates, correction_bias.inner});
-    auto orig_scores = gate_results[0];
-    auto neg_biased = gate_results[1];
-
-    auto all_indices = argpartition(neg_biased, top_k - 1, -1);
-    auto topk_indices = slice(all_indices, {0, 0}, {(int)all_indices.shape()[0], top_k});
-    auto topk_scores = take_along_axis(orig_scores, topk_indices, -1);
-
-    if (top_k > 1 && norm_topk_prob) {
-        // Compiled: normalize + scale
-        static auto normalize_fn = get_compiled_gate_normalize();
-        topk_scores = normalize_fn({topk_scores, array(scaling_factor)})[0];
-    } else {
-        topk_scores = multiply(topk_scores, array(scaling_factor));
-    }
-
-    // 2. SwitchMLP: expand + gather_qmm(fc1) + relu² + gather_qmm(fc2) + squeeze
-    auto x_shape = x.inner.shape();
-    auto x_exp = reshape(x.inner, {x_shape[0], 1, 1, x_shape[1]});
-
-    auto h = gather_qmm(
-        x_exp, fc1_weight.inner, fc1_scales.inner, fc1_biases.inner,
-        std::nullopt, topk_indices,
-        true, group_size, bits, "affine", false);
-    // relu² = relu(x)²
-    { MlxArray h_w{h}; h = compiled_relu_squared(h_w)->inner; }
-    h = gather_qmm(
-        h, fc2_weight.inner, fc2_scales.inner, fc2_biases.inner,
-        std::nullopt, topk_indices,
-        true, group_size, bits, "affine", false);
-    h = squeeze(h, -2);  // [tokens, top_k, hidden]
-
-    // 3. Score weighting: weighted sum over experts, cast back to input dtype
-    // Cast scores to h's dtype to avoid mixed float32×float16 multiply
-    // which produces NaN on M5 Max (Metal GPU Family 4) NAx broadcast kernel.
-    auto scores_exp = reshape(topk_scores, {topk_scores.shape()[0], top_k, 1});
-    auto scores_cast = astype(scores_exp, h.dtype());
-    auto result = astype(sum(multiply(h, scores_cast), -2, false), x.inner.dtype());
-
-    // 4. Optional shared expert
-    if (shared_up_weight && shared_down_weight) {
-        auto shared_h = quantized_matmul(
-            x.inner, shared_up_weight->inner, shared_up_scales->inner,
-            shared_up_biases ? std::optional(shared_up_biases->inner) : std::nullopt,
-            true, group_size, bits);
-        { MlxArray sh_w{shared_h}; shared_h = compiled_relu_squared(sh_w)->inner; }
-        shared_h = quantized_matmul(
-            shared_h, shared_down_weight->inner, shared_down_scales->inner,
-            shared_down_biases ? std::optional(shared_down_biases->inner) : std::nullopt,
-            true, group_size, bits);
-        result = add(result, shared_h);
-    }
-
-    return std::make_unique<MlxArray>(std::move(result));
-}
-
-bool ssm_kernel_available() {
-#ifdef __APPLE__
-    return mlx::core::metal::is_available();
-#else
-    return false;
-#endif
-}
-
-void ssm_update_kernel(
-    const MlxArray& hidden_states,
-    const MlxArray& A_log,
-    const MlxArray& B,
-    const MlxArray& C,
-    const MlxArray& D,
-    const MlxArray& dt,
-    const MlxArray& dt_bias,
-    const MlxArray& state_in,
-    float time_step_min,
-    float time_step_max,
-    std::unique_ptr<MlxArray>& output,
-    std::unique_ptr<MlxArray>& next_state
-) {
-    using namespace mlx::core;
-
-    auto shape = hidden_states.inner.shape();
-    int n = shape[0];  // batch
-    int h = shape[2];  // num_heads
-    int dh = shape[3]; // head_dim
-    auto b_shape = B.inner.shape();
-    int hb = b_shape[2]; // n_groups
-    int ds = b_shape[3]; // state_dim
-    int g = h / hb;      // heads per group
-
-    auto input_type = hidden_states.inner.dtype();
-    auto state_type = state_in.inner.dtype();
-
-    // Compute dt with softplus + clip (promoted to float32 internally)
-    auto dt_result = compute_dt_compiled(dt.inner, dt_bias.inner, time_step_min, time_step_max);
-
-    // Call the Metal kernel
-    auto& kernel = get_ssm_kernel().get();
-
-    // CustomKernelFunction signature:
-    // (inputs, output_shapes, output_dtypes, grid, threadgroup, template_args, init_value, verbose, stream)
-    // T = input type for x/out, U = state type for state_in/state_out (allows float32 state accumulation)
-    std::vector<std::pair<std::string, mlx::core::fast::TemplateArg>> template_args = {
-        {"T", input_type},
-        {"U", state_type},
-        {"Dh", dh},
-        {"Ds", ds},
-        {"H", h},
-        {"G", g},
-    };
-
-    std::vector<array> inputs = {
-        hidden_states.inner, A_log.inner, B.inner, C.inner,
-        mlx::core::astype(D.inner, input_type), dt_result, state_in.inner
-    };
-    std::vector<Shape> output_shapes = {Shape{n, 1, h, dh}, state_in.inner.shape()};
-    std::vector<Dtype> output_dtypes = {input_type, state_type};
-
-    auto results = kernel(
-        inputs,
-        output_shapes,
-        output_dtypes,
-        std::make_tuple(32, dh, h * n),    // grid
-        std::make_tuple(32, 8, 1),          // threadgroup
-        template_args,
-        std::nullopt,  // init_value
-        false,         // verbose
-        {}             // stream (default)
-    );
-
-    output = std::make_unique<MlxArray>(std::move(results[0]));
-    next_state = std::make_unique<MlxArray>(std::move(results[1]));
-}
-
-// ── Fused MoE expert kernel (single-token decode, power-of-2 bits) ──────────
-// Computes a decode token's routed-expert output:
-//   out[h] = sum_k scores[k] * down_k( silu(gate_k(x)) * up_k(x) )[h]
-// over the K selected experts (indices[k]), with affine-quantized gate/up/down,
-// as two Metal dispatches that stream every active expert weight once.
-//
-// A single fused launch is bandwidth-bound the wrong way at batch=1: the
-// gate/up -> down dependency goes through the per-expert activation, so keeping
-// it in threadgroup memory confines each expert to one threadgroup and caps
-// occupancy at K cores (measured ~0.46x of gather_qmm on qwen3-30b-a3b).
-// Instead we break the barrier by staging the swiglu activation in global
-// memory between two dispatches, so every GEMV output row runs as an
-// independent simdgroup across all GPU cores:
-//   A) gate/up GEMV + swiglu  -> act_g[K, Dff] (f32)
-//   B) down GEMV * score       -> partial[K, Din]
-// partial is summed over K by the host wrapper. This is non-redundant (each
-// weight read once) and beats gather_qmm by ~3.5% on qwen3-30b-a3b, greedy
-// output byte-identical. Power-of-2 bits only (4/8); callers fall back to
-// SwitchGLU for 6-bit, non-affine, or oversized configs (#268 step 2b).
-namespace {
-    // SGY = simdgroups per threadgroup; each owns one output row (grid.y) and
-    // its 32 lanes stride the contraction dim, reduced with simd_sum (cf.
-    // ssm_update_kernel). A lane unpacks a whole 32-bit pack (vpw = 32/bits
-    // weights) at a time: group_size is a multiple of vpw for bits in {4, 8},
-    // so every weight in a pack shares one (scale, bias). The reduction order
-    // differs from gather_qmm, which is fine for the RMS<5e-3 / greedy-identical
-    // gate, not bitwise parity.
-    static const char* MOE_GATEUP_METAL_SOURCE = R"(
-        uint lane  = thread_position_in_threadgroup.x;   // 0..31 (one simdgroup)
-        uint f     = thread_position_in_grid.y;          // output row 0..Dff-1
-        uint eslot = thread_position_in_grid.z;          // 0..K-1
-        if (f >= Dff) return;                            // uniform across the simdgroup
-        uint e = indices[eslot];
-
-        constexpr uint vpw   = 32u / bits;
-        constexpr uint wmask = (1u << bits) - 1u;
-        constexpr uint Din_p = Din / vpw;
-        constexpr uint G     = Din / group_size;
-
-        uint row = e * Dff + f;
-        const device uint* gwr = gate_w + row * Din_p;
-        const device T*    gsr = gate_s + row * G;
-        const device T*    gbr = gate_b + row * G;
-        const device uint* uwr = up_w   + row * Din_p;
-        const device T*    usr = up_s   + row * G;
-        const device T*    ubr = up_b   + row * G;
-        float g = 0.0f, u = 0.0f;
-        for (uint p = lane; p < Din_p; p += 32u) {
-            uint base = p * vpw;
-            uint grp  = base / group_size;
-            float gs = (float)gsr[grp], gb = (float)gbr[grp];
-            float us = (float)usr[grp], ub = (float)ubr[grp];
-            uint gpk = gwr[p];
-            uint upk = uwr[p];
-            for (uint j = 0; j < vpw; ++j) {
-                float xv = (float)x[base + j];
-                g += xv * ((float)((gpk >> (j * bits)) & wmask) * gs + gb);
-                u += xv * ((float)((upk >> (j * bits)) & wmask) * us + ub);
-            }
-        }
-        g = simd_sum(g);
-        u = simd_sum(u);
-        if (lane == 0u) {
-            act_g[eslot * Dff + f] = (g / (1.0f + fast::exp(-g))) * u;
-        }
-    )";
-
-    static const char* MOE_DOWN_METAL_SOURCE = R"(
-        uint lane  = thread_position_in_threadgroup.x;   // 0..31 (one simdgroup)
-        uint h     = thread_position_in_grid.y;          // output row 0..Din-1
-        uint eslot = thread_position_in_grid.z;          // 0..K-1
-        if (h >= Din) return;                            // uniform across the simdgroup
-        uint e = indices[eslot];
-
-        constexpr uint vpw   = 32u / bits;
-        constexpr uint wmask = (1u << bits) - 1u;
-        constexpr uint Dff_p = Dff / vpw;
-        constexpr uint Gd    = Dff / group_size;
-
-        uint row = e * Din + h;
-        const device uint*  dwr = down_w + row * Dff_p;
-        const device T*     dsr = down_s + row * Gd;
-        const device T*     dbr = down_b + row * Gd;
-        const device float* a   = act_g  + eslot * Dff;
-        float d = 0.0f;
-        for (uint p = lane; p < Dff_p; p += 32u) {
-            uint base = p * vpw;
-            uint grp  = base / group_size;
-            float ds = (float)dsr[grp], db = (float)dbr[grp];
-            uint dpk = dwr[p];
-            for (uint j = 0; j < vpw; ++j) {
-                d += a[base + j] * ((float)((dpk >> (j * bits)) & wmask) * ds + db);
-            }
-        }
-        d = simd_sum(d);
-        if (lane == 0u) {
-            out[eslot * Din + h] = (T)((float)scores[eslot] * d);
-        }
-    )";
-
-    struct MoeGateUpKernelHolder {
-        std::optional<mlx::core::fast::CustomKernelFunction> kernel;
-        bool initialized = false;
-        mlx::core::fast::CustomKernelFunction& get() {
-            if (!initialized) {
-                kernel = mlx::core::fast::metal_kernel(
-                    "moe_gateup_kernel",
-                    {"x", "indices", "gate_w", "gate_s", "gate_b",
-                     "up_w", "up_s", "up_b"},
-                    {"act_g"},
-                    MOE_GATEUP_METAL_SOURCE
-                );
-                initialized = true;
-            }
-            return *kernel;
-        }
-    };
-    static MoeGateUpKernelHolder& get_moe_gateup_kernel() {
-        static MoeGateUpKernelHolder holder;
-        return holder;
-    }
-
-    struct MoeDownKernelHolder {
-        std::optional<mlx::core::fast::CustomKernelFunction> kernel;
-        bool initialized = false;
-        mlx::core::fast::CustomKernelFunction& get() {
-            if (!initialized) {
-                kernel = mlx::core::fast::metal_kernel(
-                    "moe_down_kernel",
-                    {"indices", "down_w", "down_s", "down_b", "act_g", "scores"},
-                    {"out"},
-                    MOE_DOWN_METAL_SOURCE
-                );
-                initialized = true;
-            }
-            return *kernel;
-        }
-    };
-    static MoeDownKernelHolder& get_moe_down_kernel() {
-        static MoeDownKernelHolder holder;
-        return holder;
-    }
-}
-
-std::unique_ptr<MlxArray> fused_moe_expert_kernel(
-    const MlxArray& x,
-    const MlxArray& indices,
-    const MlxArray& gate_w, const MlxArray& gate_s, const MlxArray& gate_b,
-    const MlxArray& up_w,   const MlxArray& up_s,   const MlxArray& up_b,
-    const MlxArray& down_w, const MlxArray& down_s, const MlxArray& down_b,
-    const MlxArray& scores,
-    int32_t din, int32_t dff, int32_t k,
-    int32_t bits, int32_t group_size
-) {
-    using namespace mlx::core;
-    auto T = x.inner.dtype();
-
-    // SGY = simdgroups per threadgroup (one per output row). 8 is the measured
-    // sweet spot on qwen3-30b-a3b (M1 Ultra); env-overridable for other
-    // hardware. It feeds an MLX template arg, which JIT-specializes at runtime.
-    int sgy = 8;
-    if (const char* s = std::getenv("MLXCEL_FUSED_MOE_SGY")) {
-        int v = std::atoi(s);
-        if (v >= 1 && v <= 32) sgy = v;
-    }
-    auto round_up = [](int n, int m) { return ((n + m - 1) / m) * m; };
-    std::vector<std::pair<std::string, mlx::core::fast::TemplateArg>> ta = {
-        {"T", T}, {"K", k}, {"Din", din}, {"Dff", dff},
-        {"bits", bits}, {"group_size", group_size},
-    };
-
-    // A) gate/up + swiglu -> act_g[K, Dff] (f32 for the down GEMV).
-    auto& kA = get_moe_gateup_kernel().get();
-    std::vector<array> inA = {
-        astype(x.inner, T), astype(indices.inner, uint32),
-        gate_w.inner, astype(gate_s.inner, T), astype(gate_b.inner, T),
-        up_w.inner,   astype(up_s.inner, T),   astype(up_b.inner, T),
-    };
-    auto rA = kA(
-        inA, { Shape{k, dff} }, { float32 },
-        std::make_tuple(32, round_up(dff, sgy), k),
-        std::make_tuple(32, sgy, 1),
-        ta, std::nullopt, false, {}
-    );
-
-    // B) down * score -> partial[K, Din]; summed over K into [Din].
-    auto& kB = get_moe_down_kernel().get();
-    std::vector<array> inB = {
-        astype(indices.inner, uint32),
-        down_w.inner, astype(down_s.inner, T), astype(down_b.inner, T),
-        rA[0], astype(scores.inner, T),
-    };
-    auto rB = kB(
-        inB, { Shape{k, din} }, { T },
-        std::make_tuple(32, round_up(din, sgy), k),
-        std::make_tuple(32, sgy, 1),
-        ta, std::nullopt, false, {}
-    );
-    auto summed = sum(rB[0], /*axis=*/0, /*keepdims=*/false);
-    return std::make_unique<MlxArray>(std::move(summed));
-}
-
-// Fused Mamba2 mixer forward for single-token decode.
-// Combines in_proj + conv1d + SSM kernel + MambaRMSNormGated + out_proj into one C++ call.
-// Used by: NemotronH
-void fused_mamba2_forward(
-    const MlxArray& hidden_states,
-    const MlxArray& in_proj_weight,
-    const MlxArray& in_proj_scales,
-    const MlxArray* in_proj_biases,
-    const MlxArray& conv_weight,
-    const MlxArray* conv_bias,
-    const MlxArray& A_log,
-    const MlxArray& D,
-    const MlxArray& dt_bias,
-    const MlxArray& norm_weight,
-    const MlxArray& out_proj_weight,
-    const MlxArray& out_proj_scales,
-    const MlxArray* out_proj_biases,
-    const MlxArray& conv_state_in,
-    const MlxArray& ssm_state_in,
-    int32_t intermediate_size,
-    int32_t conv_dim,
-    int32_t conv_kernel_size,
-    int32_t num_heads,
-    int32_t head_dim,
-    int32_t n_groups,
-    int32_t ssm_state_size,
-    float time_step_min,
-    float time_step_max,
-    float norm_eps,
-    int32_t group_size,
-    int32_t bits,
-    std::unique_ptr<MlxArray>& output,
-    std::unique_ptr<MlxArray>& conv_state_out,
-    std::unique_ptr<MlxArray>& ssm_state_out
-) {
-    using namespace mlx::core;
-
-    // --- Shape extraction ---
-    auto hs_shape = hidden_states.inner.shape();
-    int batch    = (int)hs_shape[0];
-    int seq_len  = (int)hs_shape[1];   // always 1 for decode
-    int hidden   = (int)hs_shape[2];
-    int bc_size  = n_groups * ssm_state_size;
-    int proj_cols = intermediate_size + conv_dim + num_heads;
-
-    // --- Step 1: in_proj (quantized matmul, affine mode) ---
-    // Flatten to [batch*seq_len, hidden] for matmul
-    auto x_flat = reshape(hidden_states.inner, {batch * seq_len, hidden});
-    // Fast path: omit mode for affine, pass biases directly when present
-    auto projected_flat = in_proj_biases
-        ? quantized_matmul(x_flat, in_proj_weight.inner, in_proj_scales.inner,
-                          in_proj_biases->inner, true, group_size, bits)
-        : quantized_matmul(x_flat, in_proj_weight.inner, in_proj_scales.inner,
-                          std::nullopt, true, group_size, bits);
-    // Restore to [batch, seq_len, proj_cols]
-    auto projected = reshape(projected_flat, {batch, seq_len, proj_cols});
-
-    // Dtype trace (temporary)
-    if (std::getenv("MLXCEL_TRACE_DTYPE")) {
-        mlx::core::eval(projected);
-        std::cerr << "[C++ DTYPE] after in_proj: " << projected.dtype() << std::endl;
-    }
-    // --- Step 2: Split projected into gate, conv_input, dt ---
-    // gate:       [batch, seq_len, intermediate_size]
-    // conv_input: [batch, seq_len, conv_dim]
-    // dt:         [batch, seq_len, num_heads]
-    auto gate       = slice(projected,
-                            {0, 0, 0},
-                            {batch, seq_len, intermediate_size});
-    auto conv_input = slice(projected,
-                            {0, 0, intermediate_size},
-                            {batch, seq_len, intermediate_size + conv_dim});
-    auto dt_raw     = slice(projected,
-                            {0, 0, intermediate_size + conv_dim},
-                            {batch, seq_len, proj_cols});
-
-    // --- Step 3: Depthwise conv1d with sliding state ---
-    // Concatenate conv_state_in [batch, k-1, conv_dim] with conv_input [batch, 1, conv_dim]
-    // -> padded_input [batch, k, conv_dim]  (k = conv_kernel_size)
-    auto padded_input = concatenate(std::vector<array>{conv_state_in.inner, conv_input}, 1);
-
-    // New conv state: last (k-1) elements of padded_input on axis 1
-    int padded_len      = conv_kernel_size - 1 + seq_len;
-    int new_state_start = padded_len - (conv_kernel_size - 1);
-    auto new_conv_state = slice(padded_input,
-                                {0, new_state_start, 0},
-                                {batch, padded_len, conv_dim});
-
-    // Depthwise conv1d: stride=1, padding=0, dilation=1, groups=conv_dim
-    auto conv_out = mlx::core::conv1d(
-        padded_input, conv_weight.inner,
-        /*stride=*/1, /*padding=*/0, /*dilation=*/1, /*groups=*/conv_dim);
-
-    // Optional bias: reshape [conv_dim] -> [1, 1, conv_dim] for broadcast
-    if (conv_bias) {
-        auto bias_r = reshape(conv_bias->inner, {1, 1, conv_dim});
-        conv_out = add(conv_out, bias_r);
-    }
-
-    // SiLU: x * sigmoid(x) — compiled kernel fusion
-    MlxArray co_w{conv_out};
-    auto conv_output = compiled_silu(co_w)->inner;
-
-    // --- Step 4: Split conv_output into hidden_ssm, B, C ---
-    // conv_output: [batch, seq_len, conv_dim]
-    // hidden_ssm:  [batch, seq_len, intermediate_size]
-    // B, C:        [batch, seq_len, n_groups * ssm_state_size] each
-    auto hidden_ssm = slice(conv_output,
-                            {0, 0, 0},
-                            {batch, seq_len, intermediate_size});
-    auto B          = slice(conv_output,
-                            {0, 0, intermediate_size},
-                            {batch, seq_len, intermediate_size + bc_size});
-    auto C          = slice(conv_output,
-                            {0, 0, intermediate_size + bc_size},
-                            {batch, seq_len, conv_dim});
-
-    // --- Step 5: Reshape inputs for SSM kernel ---
-    // x:  [batch, seq_len, num_heads, head_dim]
-    // B:  [batch, seq_len, n_groups, ssm_state_size]
-    // C:  [batch, seq_len, n_groups, ssm_state_size]
-    // dt: [batch, seq_len, num_heads]   (already the right last dim)
-    auto x_ssm = reshape(hidden_ssm, {batch, seq_len, num_heads, head_dim});
-    auto B_r    = reshape(B,         {batch, seq_len, n_groups,  ssm_state_size});
-    auto C_r    = reshape(C,         {batch, seq_len, n_groups,  ssm_state_size});
-
-    // --- Step 6: Fused SSM Metal kernel ---
-    // Wrap plain arrays in MlxArray for the kernel call
-    MlxArray x_ssm_w{x_ssm};
-    MlxArray B_w{B_r};
-    MlxArray C_w{C_r};
-    MlxArray dt_w{dt_raw};
-
-    std::unique_ptr<MlxArray> ssm_y;
-    std::unique_ptr<MlxArray> new_ssm_state;
-
-    // ssm_update_kernel is defined above in this translation unit
-    ssm_update_kernel(
-        x_ssm_w, A_log, B_w, C_w, D, dt_w, dt_bias,
-        ssm_state_in,
-        time_step_min, time_step_max,
-        ssm_y, new_ssm_state);
-
-    // ssm_y: [batch, 1, num_heads, head_dim] -> [batch, seq_len, intermediate_size]
-    if (std::getenv("MLXCEL_TRACE_DTYPE")) {
-        mlx::core::eval(ssm_y->inner);
-        std::cerr << "[C++ DTYPE] ssm_y: " << ssm_y->inner.dtype() << std::endl;
-    }
-    auto y = reshape(ssm_y->inner, {batch, seq_len, intermediate_size});
-
-    // --- Step 7: MambaRMSNormGated ---
-    // 7a. Gate: y_gated = y * silu(gate)
-    // gate_activated = silu(gate), then y_gated = y * gate_activated
-    MlxArray gate_w{gate};
-    auto y_gated = multiply(y, compiled_silu(gate_w)->inner);
-
-    // 7b. Grouped RMS norm: reshape last dim into [batch, seq_len, n_groups, group_size]
-    //     where group_size = intermediate_size / n_groups (NOT head_dim!)
-    int norm_group_size = intermediate_size / n_groups;
-    auto y_grouped = reshape(y_gated, {batch, seq_len, n_groups, norm_group_size});
-    auto unit_weight = mlx::core::ones({norm_group_size}, hidden_states.inner.dtype());
-    auto y_normed_grouped = mlx::core::fast::rms_norm(y_grouped, unit_weight, norm_eps);
-
-    // 7c. Flatten back and apply learned norm weight
-    auto y_normed_flat = reshape(y_normed_grouped, {batch, seq_len, intermediate_size});
-    auto y_normed = multiply(norm_weight.inner, y_normed_flat);
-
-    // --- Step 8: out_proj (quantized matmul, affine mode) ---
-    auto y_proj_flat = reshape(y_normed, {batch * seq_len, intermediate_size});
-    auto out_flat = out_proj_biases
-        ? quantized_matmul(y_proj_flat, out_proj_weight.inner, out_proj_scales.inner,
-                          out_proj_biases->inner, true, group_size, bits)
-        : quantized_matmul(y_proj_flat, out_proj_weight.inner, out_proj_scales.inner,
-                          std::nullopt, true, group_size, bits);
-    auto out = reshape(out_flat, {batch, seq_len, hidden});
-    if (std::getenv("MLXCEL_TRACE_DTYPE")) {
-        mlx::core::eval(out);
-        std::cerr << "[C++ DTYPE] final out: " << out.dtype() << std::endl;
-    }
-
-    // --- Write outputs ---
-    output         = std::make_unique<MlxArray>(std::move(out));
-    conv_state_out = std::make_unique<MlxArray>(std::move(new_conv_state));
-    ssm_state_out  = std::move(new_ssm_state);
-}
-
-// ============ NemotronH Full Forward Decode ============
-
-namespace {
-struct NemotronModel {
-    // Global weights
-    const MlxArray *embed_w, *embed_s, *embed_b;
-    const MlxArray *final_norm_w;
-    const MlxArray *lm_head_w, *lm_head_s, *lm_head_b;
-    std::vector<const MlxArray*> norm_weights;
-    std::vector<int32_t> block_types;
-
-    struct Mamba {
-        const MlxArray *in_w, *in_s, *in_b;
-        const MlxArray *conv_w, *conv_b;
-        const MlxArray *a_log, *d, *dt_bias, *norm_w;
-        const MlxArray *out_w, *out_s, *out_b;
-    };
-    struct MoE {
-        const MlxArray *gate_w, *corr_bias;
-        const MlxArray *fc1_w, *fc1_s, *fc1_b;
-        const MlxArray *fc2_w, *fc2_s, *fc2_b;
-        const MlxArray *su_w, *su_s, *su_b;
-        const MlxArray *sd_w, *sd_s, *sd_b;
-    };
-    struct Attn {
-        const MlxArray *q_w, *q_s, *q_b;
-        const MlxArray *k_w, *k_s, *k_b;
-        const MlxArray *v_w, *v_s, *v_b;
-        const MlxArray *o_w, *o_s, *o_b;
-    };
-    std::vector<Mamba> mamba_layers;
-    std::vector<MoE> moe_layers;
-    std::vector<Attn> attn_layers;
-
-    // Config
-    float norm_eps; int gs, bits;
-    int m_inter, m_cdim, m_ck, m_heads, m_hdim, m_groups, m_state;
-    float m_ts_min, m_ts_max, m_neps;
-    int moe_tk; float moe_sc; bool moe_norm;
-    int a_heads, a_kvh, a_hdim; float a_rope, a_scale;
-};
-
-static std::unordered_map<uint64_t, std::unique_ptr<NemotronModel>> g_models;
-static uint64_t g_next_handle = 1;
-} // anonymous namespace
-
-uint64_t nemotron_register_model(
-    const MlxArray& embed_w, const MlxArray& embed_s, const MlxArray& embed_b,
-    const MlxArray& final_norm_w,
-    const MlxArray& lm_head_w, const MlxArray& lm_head_s, const MlxArray* lm_head_b,
-    rust::Slice<const MlxArray* const> norm_weights,
-    rust::Slice<const int32_t> block_types,
-    rust::Slice<const MlxArray* const> mw,
-    rust::Slice<const MlxArray* const> ew,
-    rust::Slice<const MlxArray* const> aw,
-    float norm_eps, int32_t gs, int32_t bits,
-    int32_t m_inter, int32_t m_cdim, int32_t m_ck,
-    int32_t m_heads, int32_t m_hdim, int32_t m_groups, int32_t m_state,
-    float m_ts_min, float m_ts_max, float m_neps,
-    int32_t moe_tk, float moe_sc, bool moe_norm,
-    int32_t a_heads, int32_t a_kvh, int32_t a_hdim,
-    float a_rope, float a_scale
-) {
-    auto model = std::make_unique<NemotronModel>();
-    model->embed_w = &embed_w; model->embed_s = &embed_s; model->embed_b = &embed_b;
-    model->final_norm_w = &final_norm_w;
-    model->lm_head_w = &lm_head_w; model->lm_head_s = &lm_head_s;
-    model->lm_head_b = lm_head_b;
-    for (auto* p : norm_weights) model->norm_weights.push_back(p);
-    for (auto bt : block_types) model->block_types.push_back(bt);
-
-    // Parse Mamba weights: 12 per layer
-    // Order: in_w, in_s, in_b, conv_w, conv_b, a_log, d, dt_bias, norm_w, out_w, out_s, out_b
-    for (size_t i = 0; i + 11 < mw.size(); i += 12) {
-        model->mamba_layers.push_back({
-            mw[i], mw[i+1], mw[i+2],  // in(w,s,b)
-            mw[i+3], mw[i+4],          // conv(w,b)
-            mw[i+5], mw[i+6], mw[i+7], mw[i+8],  // a_log, d, dt_bias, norm_w
-            mw[i+9], mw[i+10], mw[i+11] // out(w,s,b)
-        });
-    }
-    // Parse MoE weights: 14 per layer
-    for (size_t i = 0; i + 13 < ew.size(); i += 14) {
-        model->moe_layers.push_back({
-            ew[i], ew[i+1],
-            ew[i+2], ew[i+3], ew[i+4],
-            ew[i+5], ew[i+6], ew[i+7],
-            ew[i+8], ew[i+9], ew[i+10],
-            ew[i+11], ew[i+12], ew[i+13]
-        });
-    }
-    // Parse Attention weights: 12 per layer
-    for (size_t i = 0; i + 11 < aw.size(); i += 12) {
-        model->attn_layers.push_back({
-            aw[i], aw[i+1], aw[i+2],
-            aw[i+3], aw[i+4], aw[i+5],
-            aw[i+6], aw[i+7], aw[i+8],
-            aw[i+9], aw[i+10], aw[i+11]
-        });
-    }
-
-    model->norm_eps = norm_eps; model->gs = gs; model->bits = bits;
-    model->m_inter = m_inter; model->m_cdim = m_cdim; model->m_ck = m_ck;
-    model->m_heads = m_heads; model->m_hdim = m_hdim; model->m_groups = m_groups;
-    model->m_state = m_state; model->m_ts_min = m_ts_min; model->m_ts_max = m_ts_max;
-    model->m_neps = m_neps;
-    model->moe_tk = moe_tk; model->moe_sc = moe_sc; model->moe_norm = moe_norm;
-    model->a_heads = a_heads; model->a_kvh = a_kvh; model->a_hdim = a_hdim;
-    model->a_rope = a_rope; model->a_scale = a_scale;
-
-    uint64_t handle = g_next_handle++;
-    g_models[handle] = std::move(model);
-    return handle;
-}
-
-void nemotron_free_model(uint64_t handle) {
-    g_models.erase(handle);
-}
-
-void nemotron_decode_step(
-    uint64_t handle,
-    const MlxArray& input_ids,
-    rust::Slice<const MlxArray* const> mamba_conv_in,
-    rust::Slice<const MlxArray* const> mamba_ssm_in,
-    rust::Slice<const MlxArray* const> attn_kv_keys,
-    rust::Slice<const MlxArray* const> attn_kv_values,
-    rust::Slice<const int32_t> attn_kv_offsets,
-    std::unique_ptr<MlxArray>& logits,
-    rust::Slice<std::unique_ptr<MlxArray>> mamba_conv_out,
-    rust::Slice<std::unique_ptr<MlxArray>> mamba_ssm_out
-) {
-    using namespace mlx::core;
-    auto& m = *g_models.at(handle);
-    int num_layers = (int)m.block_types.size();
-
-    // Embedding (quantized)
-    auto flat_ids = reshape(input_ids.inner, {-1});
-    auto w_idx = take(m.embed_w->inner, flat_ids, 0);
-    auto s_idx = take(m.embed_s->inner, flat_ids, 0);
-    auto b_idx = take(m.embed_b->inner, flat_ids, 0);
-    auto h = dequantize(w_idx, s_idx, b_idx, m.gs, m.bits, "affine");
-    auto id_shape = input_ids.inner.shape();
-    h = reshape(h, {id_shape[0], id_shape[1], (int)h.shape().back()});
-
-    int mamba_idx = 0, moe_idx = 0, attn_idx = 0;
-
-    for (int i = 0; i < num_layers; ++i) {
-        // RMSNorm
-        auto normed = fast::rms_norm(h, m.norm_weights[i]->inner, m.norm_eps);
-
-        int bt = m.block_types[i];
-        array out = normed; // placeholder
-
-        if (bt == 0) { // Mamba
-            auto& ml = m.mamba_layers[mamba_idx];
-            MlxArray h_w{normed};
-            MlxArray in_w{ml.in_w->inner}, in_s{ml.in_s->inner};
-            MlxArray cw{ml.conv_w->inner}, al{ml.a_log->inner}, dd{ml.d->inner}, dtb{ml.dt_bias->inner};
-            MlxArray nw{ml.norm_w->inner}, ow{ml.out_w->inner}, os{ml.out_s->inner};
-            MlxArray cs{mamba_conv_in[mamba_idx]->inner}, ss{mamba_ssm_in[mamba_idx]->inner};
-
-            std::unique_ptr<MlxArray> m_out, m_cs, m_ss;
-            fused_mamba2_forward(
-                h_w, in_w, in_s, ml.in_b, cw, ml.conv_b,
-                al, dd, dtb, nw, ow, os, ml.out_b,
-                cs, ss,
-                m.m_inter, m.m_cdim, m.m_ck,
-                m.m_heads, m.m_hdim, m.m_groups, m.m_state,
-                m.m_ts_min, m.m_ts_max, m.m_neps,
-                m.gs, m.bits,
-                m_out, m_cs, m_ss
-            );
-            out = m_out->inner;
-            mamba_conv_out[mamba_idx] = std::move(m_cs);
-            mamba_ssm_out[mamba_idx] = std::move(m_ss);
-            mamba_idx++;
-        } else if (bt == 3) { // MoE
-            auto& el = m.moe_layers[moe_idx];
-            // Flatten to [batch*seq, hidden] for MoE (expects 2D input)
-            auto normed_flat = reshape(normed, {-1, (int)normed.shape().back()});
-            auto moe_result = fused_moe_forward(
-                MlxArray{normed_flat},
-                *el.gate_w, *el.corr_bias,
-                *el.fc1_w, *el.fc1_s, *el.fc1_b,
-                *el.fc2_w, *el.fc2_s, *el.fc2_b,
-                el.su_w, el.su_s, el.su_b,
-                el.sd_w, el.sd_s, el.sd_b,
-                m.moe_tk, m.moe_sc, m.moe_norm,
-                m.gs, m.bits
-            );
-            // Reshape back to [batch, seq, hidden]
-            out = reshape(moe_result->inner, h.shape());
-            moe_idx++;
-        } else if (bt == 1) { // Attention
-            auto& al = m.attn_layers[attn_idx];
-            int batch = (int)h.shape()[0];
-            int seq = (int)h.shape()[1];
-            int hidden = (int)h.shape()[2];
-            auto x_flat = reshape(normed, {batch * seq, hidden});
-
-            // QKV projections
-            std::optional<array> qb = al.q_b ? std::optional(al.q_b->inner) : std::nullopt;
-            std::optional<array> kb = al.k_b ? std::optional(al.k_b->inner) : std::nullopt;
-            std::optional<array> vb = al.v_b ? std::optional(al.v_b->inner) : std::nullopt;
-            std::optional<array> ob = al.o_b ? std::optional(al.o_b->inner) : std::nullopt;
-            auto q = quantized_matmul(x_flat, al.q_w->inner, al.q_s->inner, qb, true, m.gs, m.bits);
-            auto k = quantized_matmul(x_flat, al.k_w->inner, al.k_s->inner, kb, true, m.gs, m.bits);
-            auto v = quantized_matmul(x_flat, al.v_w->inner, al.v_s->inner, vb, true, m.gs, m.bits);
-
-            q = reshape(q, {batch, seq, m.a_heads, m.a_hdim});
-            k = reshape(k, {batch, seq, m.a_kvh, m.a_hdim});
-            v = reshape(v, {batch, seq, m.a_kvh, m.a_hdim});
-            q = transpose(q, {0, 2, 1, 3});
-            k = transpose(k, {0, 2, 1, 3});
-            v = transpose(v, {0, 2, 1, 3});
-
-            int offset = attn_kv_offsets[attn_idx];
-            q = fast::rope(q, m.a_hdim, false, m.a_rope, 1.0f, offset);
-            k = fast::rope(k, m.a_hdim, false, m.a_rope, 1.0f, offset);
-
-            // Simple KV cache: concatenate with existing
-            if (attn_kv_keys[attn_idx]) {
-                k = concatenate(std::vector<array>{attn_kv_keys[attn_idx]->inner, k}, 2);
-                v = concatenate(std::vector<array>{attn_kv_values[attn_idx]->inner, v}, 2);
-            }
-
-            // GQA repeat
-            int n_rep = m.a_heads / m.a_kvh;
-            if (n_rep > 1) {
-                auto ks = k.shape();
-                k = reshape(k, {ks[0], ks[1], 1, ks[2], ks[3]});
-                k = broadcast_to(k, {ks[0], ks[1], n_rep, ks[2], ks[3]});
-                k = reshape(k, {ks[0], ks[1] * n_rep, ks[2], ks[3]});
-                auto vs = v.shape();
-                v = reshape(v, {vs[0], vs[1], 1, vs[2], vs[3]});
-                v = broadcast_to(v, {vs[0], vs[1], n_rep, vs[2], vs[3]});
-                v = reshape(v, {vs[0], vs[1] * n_rep, vs[2], vs[3]});
-            }
-
-            auto attn = fast::scaled_dot_product_attention(q, k, v, m.a_scale);
-            attn = transpose(attn, {0, 2, 1, 3});
-            attn = reshape(attn, {batch, seq, m.a_heads * m.a_hdim});
-            auto o_flat = reshape(attn, {batch * seq, m.a_heads * m.a_hdim});
-            out = reshape(quantized_matmul(o_flat, al.o_w->inner, al.o_s->inner, ob, true, m.gs, m.bits),
-                         {batch, seq, hidden});
-            attn_idx++;
-        }
-
-        // Residual
-        h = add(h, out);
-    }
-
-    // Final norm + lm_head
-    h = fast::rms_norm(h, m.final_norm_w->inner, m.norm_eps);
-    auto h_flat = reshape(h, {(int)h.shape()[0], (int)h.shape().back()});
-    std::optional<array> lm_b = m.lm_head_b ? std::optional(m.lm_head_b->inner) : std::nullopt;
-    auto lm_out = quantized_matmul(h_flat, m.lm_head_w->inner, m.lm_head_s->inner, lm_b, true, m.gs, m.bits);
-    lm_out = reshape(lm_out, {(int)h.shape()[0], 1, (int)lm_out.shape().back()});
-
-    logits = std::make_unique<MlxArray>(std::move(lm_out));
-}
-
-// ── Native safetensors loading ────────────────────────────────────────
-
-// ── Metal 4 fused attention kernel scaffolding ────────────────────────────────
-//
-// BACKGROUND: Metal 4 (available on M5 and later) introduces TensorOps —
-// first-class tensor primitives built into the Metal Shading Language (MSL).
-// Key capabilities relevant to attention fusion:
-//
-//   - MTLTensor: Native GPU tensor resource type with explicit layout control.
-//     Allows keeping intermediate Q, K, V, scores, and context vectors
-//     on-chip in registers between shader stages.
-//
-//   - Tensor matmul / reduction: MSL built-in operators for fusing QKV
-//     projection, RoPE, score computation, softmax, and output projection
-//     into a SINGLE GPU dispatch, eliminating all intermediate memory writes.
-//
-//   - MTL4MachineLearningCommandEncoder: Encodes full ML subgraphs onto the
-//     GPU command timeline, enabling the Neural Accelerator to execute an
-//     entire attention layer without returning to the CPU scheduler.
-//
-// CURRENT STATUS:
-//   - For standard SDPA, this bridge delegates to upstream MLX
-//     `fast::scaled_dot_product_attention()`, which can dispatch to the
-//     backend Metal/NAX fused kernels when eligible.
-//   - For softcap attention (Gemma family), this bridge currently keeps the
-//     existing compiled softcap fallback.
-//   - window_size metadata is currently represented by explicit masks from
-//     Rust model code (no dedicated on-chip band-mask kernel here yet).
-//
-// WHEN TO IMPLEMENT THE FULL KERNEL:
-//   Requirements:
-//     - macOS 26.2 or later (first macOS release supporting Metal 4)
-//     - Xcode with Metal 4 SDK (WWDC25 release cycle)
-//     - M5 hardware for compilation and testing
-//   Reference material:
-//     - WWDC25 "Metal 4 TensorOps" session
-//     - WWDC25 "Accelerate ML inference with Metal 4" session
-//     - https://github.com/liuliu/example_matmul_metal4  (open-source matmul example)
-//     - mlx/backend/metal/steel_attention.metal  (MLX baseline for reference)
-//
-std::unique_ptr<MlxArray> fused_metal4_attention(
-    const MlxArray& q,
-    const MlxArray& k,
-    const MlxArray& v,
-    float scale,
-    const MlxArray* mask,
-    float softcap,
-    int32_t window_size,
-    bool use_metal4
-) {
-    // window_size is currently represented by an explicit mask from Rust.
-    (void)window_size;
-
-    std::optional<mlx::core::array> mask_opt = std::nullopt;
-    std::string mask_mode = "";
-    if (mask) {
-        auto m = mask->inner;
-        // Float masks are additive masks and should match q dtype.
-        // Bool/int masks must preserve original semantics.
-        if (mlx::core::issubdtype(m.dtype(), mlx::core::floating) &&
-            m.dtype() != q.inner.dtype()) {
-            m = mlx::core::astype(m, q.inner.dtype());
-        }
-        mask_opt = m;
-        mask_mode = "array";
-    }
-
-    // Keep existing softcap semantics (including GQA) on the Metal4 bridge path
-    // until upstream MLX provides a native fused softcap SDPA variant.
-    if (softcap > 0.0f) {
-        auto q_heads = q.inner.shape(1);
-        auto kv_heads = k.inner.shape(1);
-        std::unique_ptr<MlxArray> mask_holder = nullptr;
-        if (mask_opt.has_value()) {
-            mask_holder = std::make_unique<MlxArray>(mask_opt.value());
-        }
-        const MlxArray* mask_ptr = mask_holder ? mask_holder.get() : nullptr;
-
-        if (q_heads > kv_heads && q_heads % kv_heads == 0) {
-            int32_t n_rep = static_cast<int32_t>(q_heads / kv_heads);
-            return compiled_softcap_sdpa_gqa(q, k, v, scale, softcap, n_rep, mask_ptr);
-        }
-        return compiled_softcap_sdpa(q, k, v, scale, softcap, mask_ptr);
-    }
-
-    // Standard path: delegate to upstream MLX fast SDPA so backend kernel
-    // selection (including M5 NAX full/vector kernels) remains active.
-    (void)use_metal4;
-    return std::make_unique<MlxArray>(mlx::core::fast::scaled_dot_product_attention(
-        q.inner, k.inner, v.inner, scale, mask_mode, mask_opt
-    ));
-}
-
-// Fused Sparse-V SDPA Metal kernel launcher. Implementation in
-// `src/lib/mlx-cpp/turbo/sparse_v_sdpa.cpp`; we forward the call here so the
-// new symbol shows up in the cxx-bridge ABI without bloating the bridge .cpp.
-std::unique_ptr<MlxArray> turbo_sparse_v_weighted_sum(
-    const MlxArray& attn_weights,
-    const MlxArray& v_packed,
-    const MlxArray& v_rescale,
-    const MlxArray& codebook,
-    int32_t dim,
-    int32_t n_rep,
-    float threshold) {
-    auto out = mlxcel::turbo::sparse_v_weighted_sum(
-        attn_weights.inner,
-        v_packed.inner,
-        v_rescale.inner,
-        codebook.inner,
-        dim,
-        n_rep,
-        threshold);
-    return std::make_unique<MlxArray>(std::move(out));
-}
-
-// Fused Turbo4Delegated cold-V weighted-sum kernel launcher.
-// Implementation in `src/lib/mlx-cpp/turbo/turbo4_delegated_sdpa.cpp`; we
-// forward the call here so the new symbol shows up in the cxx-bridge ABI
-// without bloating the bridge .cpp. The unrotated cold weighted sum returned
-// here is paired with a host-side hot V matmul; the dequantised cold V never
-// materialises in global memory (the whole point).
-std::unique_ptr<MlxArray> turbo4_delegated_cold_weighted_sum(
-    const MlxArray& attn_weights_cold,
-    const MlxArray& v_packed_cold,
-    const MlxArray& v_rescale_cold,
-    const MlxArray& codebook,
-    int32_t dim,
-    int32_t n_rep,
-    float threshold) {
-    auto out = mlxcel::turbo::turbo4_delegated_cold_weighted_sum(
-        attn_weights_cold.inner,
-        v_packed_cold.inner,
-        v_rescale_cold.inner,
-        codebook.inner,
-        dim,
-        n_rep,
-        threshold);
-    return std::make_unique<MlxArray>(std::move(out));
-}
-
-std::unique_ptr<MlxArray> turbo4_delegated_bulk_dequant_rotated(
-    const MlxArray& v_packed,
-    const MlxArray& v_rescale,
-    const MlxArray& codebook,
-    int32_t dim) {
-    auto out = mlxcel::turbo::turbo4_delegated_bulk_dequant_rotated(
-        v_packed.inner,
-        v_rescale.inner,
-        codebook.inner,
-        dim);
-    return std::make_unique<MlxArray>(std::move(out));
-}
-
-// Steel-attention-envelope fused Turbo4Delegated SDPA bridge.
-// Wraps the launcher in `src/lib/mlx-cpp/turbo/turbo4_delegated_sdpa.cpp` and
-// repackages the std::vector<mlx::core::array> return into the `cxx`-friendly
-// `Turbo4DelegatedSteelOutputs` struct (cxx does not directly model multiple
-// return values, but it does model unique_ptr<struct>).
-std::unique_ptr<Turbo4DelegatedSteelOutputs> turbo4_delegated_steel_sdpa(
-    const MlxArray& scores,
-    const MlxArray& cold_packed,
-    const MlxArray& cold_rescale,
-    const MlxArray& hot_v,
-    const MlxArray& codebook,
-    int32_t dim,
-    int32_t n_rep,
-    int32_t cold_offset,
-    int32_t hot_offset,
-    float threshold) {
-    auto outs = mlxcel::turbo::turbo4_delegated_steel_sdpa(
-        scores.inner,
-        cold_packed.inner,
-        cold_rescale.inner,
-        hot_v.inner,
-        codebook.inner,
-        dim,
-        n_rep,
-        cold_offset,
-        hot_offset,
-        threshold);
-    auto wrapper = std::make_unique<Turbo4DelegatedSteelOutputs>();
-    wrapper->out_cold_pre = std::make_unique<MlxArray>(std::move(outs[0]));
-    wrapper->out_hot      = std::make_unique<MlxArray>(std::move(outs[1]));
-    return wrapper;
-}
-
-// Fused paged-attention decode kernel launcher (epic #116 Phase 6, #123).
-// Implementation in `src/lib/mlx-cpp/turbo/paged_attention.cpp`; forwarded here
-// so the new symbol shows up in the cxx-bridge ABI. Reads scattered KV blocks
-// out of the global pool via the block table with no separate gather copy.
-std::unique_ptr<MlxArray> paged_attention_decode(
-    const MlxArray& q,
-    const MlxArray& k_pool,
-    const MlxArray& v_pool,
-    const MlxArray& rows,
-    const MlxArray& row_offsets,
-    const MlxArray& logical_starts,
-    const MlxArray& visible_lens,
-    float scale) {
-    auto out = mlxcel::turbo::paged_attention_decode(
-        q.inner,
-        k_pool.inner,
-        v_pool.inner,
-        rows.inner,
-        row_offsets.inner,
-        logical_starts.inner,
-        visible_lens.inner,
-        scale);
-    return std::make_unique<MlxArray>(std::move(out));
-}
-
-std::unique_ptr<MlxLoadedWeights> mlx_load_safetensors(rust::Str path) {
-    std::string path_str(path.data(), path.size());
-    auto [weights_map, metadata] = mlx::core::load_safetensors(path_str);
-
-    auto result = std::make_unique<MlxLoadedWeights>();
-    result->names.reserve(weights_map.size());
-    result->arrays.reserve(weights_map.size());
-
-    for (auto& [name, arr] : weights_map) {
-        result->names.push_back(std::move(name));
-        result->arrays.push_back(std::make_unique<MlxArray>(std::move(arr)));
-    }
-
-    return result;
-}
-
-size_t loaded_weights_len(const MlxLoadedWeights& w) {
-    return w.names.size();
-}
-
-rust::String loaded_weights_name(const MlxLoadedWeights& w, size_t index) {
-    return rust::String(w.names.at(index));
-}
-
-std::unique_ptr<MlxArray> loaded_weights_take(MlxLoadedWeights& w, size_t index) {
-    return std::move(w.arrays.at(index));
-}
-
-// steel SDPA output takers. The cxx bridge does not directly
-// model destructuring a struct returned over the FFI boundary; we expose two
-// simple `move-out` accessors that the Rust caller invokes once each. After
-// both calls the struct's `unique_ptr` slots are empty (a second call would
-// return a null `unique_ptr`, which is harmless because the Rust side drops
-// the struct after the second take).
-std::unique_ptr<MlxArray> steel_outputs_take_cold(Turbo4DelegatedSteelOutputs& o) {
-    return std::move(o.out_cold_pre);
-}
-
-std::unique_ptr<MlxArray> steel_outputs_take_hot(Turbo4DelegatedSteelOutputs& o) {
-    return std::move(o.out_hot);
-}
-
-} // namespace mlx_cxx
+}  // namespace mlx_cxx
